Merged PR 985: Static images will be served from the server
Static images like the purple banner image will now hosted and served from the server itself.
Note that to run the local server you must do `gulp ngrok-serve` and not `gulp serve` & `gulp start-ngrok` for the image to be served properly. This is because the `process.env.HOSTNAME` variable is updated with the latest ngrok url. In production we'll probably set the `env.HOSTNAME` manually.