Merged PR 946: Changed database schema files to add interfaces for strong typing of database documents.
Added interfaces to AMASession.ts, Question.ts, and User.ts for strong typing of variables and function signatures which use database documents.

New Exposed Interfaces:
* IAMASession: The main interface for an AMASession document. 'hostId' field is a string Id referencing the User document of the AMA host/creator.
* IAMASession_populated: The interface used when the hostId field of the AMASession document is populated.
* IQuestion: The main interface for a Question document. 'userId' and 'amaSessionId' fields are string references to the User document of the question creator, and the AMASession document the question is linked to. The 'voters' field stores an array of string references to User documents of the users who upvoted that question.
* IQuestionPopulatedUser: The interface used when only the userId field is populated, but amaSessionId and voters fields remain as references.
* IUser: The main interface for a User document.

Be aware that typecasting from a main interface to a populated interface will be required when populating a field, as mongoose doesn't do the type conversion itself. As a result, type guards will need to be written. An example of this is provided in Database.ts in my PR merging shayan/upvoteLeaderboard into development.