Merged PR 947: Upvoting Feature
Implemented upvoting feature.

Feature breakdown:
* Users can upvote other users' questions, but not their own.
* Users cannot remove their upvote from a question. Will have to discuss if we want this feature or not.
* The leaderboard will be refreshed upon upvoting to reflect the latest within the database.

Tests:
* Existing user in database upvotes a question they have not yet upvoted.
* Existing user in database upvotes the same question twice.
* User who is not yet in the database upvotes a question.

Also added some of the new database schema types to certain variables and function signatures. Will create another PR to strongly type the remaining database variable / function signatures after the PRs for the base features of the bot have been merged.

An example for a type guard for a populate operation can be seen on line 129 of Database.ts. This type guard is used on line 118 of Database.ts.