Merged PR 1025: Revised End AMA Feature
The "End AMA" button has now moved to the leaderboard, and will only be shown if the user is the host and the AMA is still active. Otherwise, the same leaderboard with no button is shown for everyone.

Files Involved:
Questionly.ts
Controller.ts
Database.ts
Leaderboard.ts

Please disregard the error handling- the PR for App Insights addresses this.

Related work items: #10958