Merged PR 1061: activity handler and master card tests
Tests for the master card and activity handler. In the activity handler tests, for each handler I tested at least one path and for most handlers I tested a couple different paths. One thing I mainly tested was whether the data embedded in the master card (such as the session id, aadObject id) was accessed properly or not. There's no types for that so it's easy to do it incorrectly. The tests will catch these errors.